edsv: support for SAF SaTimeT timeout variable in MDS transport [#1658] Issue : In current implementation their is a mismatch of variable sizes between SAF API accepting SaTimeT variable type for timeout ( 64 bit) and MDS transport implementation accepting (32 bit) timeout variable , the MDS transport should adopt the size of the timeout parameter from 32 bits to 64 bits.
Fix :
Now MDS provide interface to accept SaTimeT type timeout variable from SAF APIS

+++ osaf/services/saf/edsv/eds/eds_ckpt.c
@@ -827,7 +827,7 @@ uint32_t eds_edu_enc_reten_list(EDS_CB *cb, NCS_UBAID *uba)
 	uint32_t rc = NCSCC_RC_SUCCESS, num_rec = 0;
 	uint8_t *pheader = NULL;
 	EDS_CKPT_HEADER ckpt_hdr;
-	uint32_t remaining_time = 0;
+	int64_t remaining_time = 0;
 	SaUint8T list_iter;
 	EDS_WORKLIST *wp = NULL;
 	TRACE_ENTER();
